requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
scikit-learn==1.5.0
joblib==1.4.2
pydantic[email]==2.8.0
paramiko==3.5.0
cryptography==42.0.5
python-json-logger==2.0.7
puttykeys==1.0.3
